CRUD post y  plan

+++ routes/web.php
@@ -27,14 +27,20 @@ Route::get('single-post/{id}', [PostController::class, 'viewSinglePost'])->name(
 
 Route::get('admin', [AdminController::class, 'index'])->middleware('auth')->name('admin.index');
 
-Route::post('admin/category/image', [CategoryController::class, 'addImage'])->name('category.addImage');
+Route::post('admin/category/image', [CategoryController::class, 'addImage'])->middleware('auth')->name('category.addImage');
 
-Route::delete('admin/category/image/{image}', [CategoryController::class, 'deleteImage'])->name('category.deleteImage');
+Route::delete('admin/category/image/{image}', [CategoryController::class, 'deleteImage'])->middleware('auth')->name('category.deleteImage');
 
-Route::resource('admin/category', CategoryController::class)->names('category');
+Route::resource('admin/category', CategoryController::class)->middleware('auth')->names('category');
 
-Route::resource('admin/post', PostController::class)->names('post');
+Route::resource('admin/post', PostController::class)->middleware('auth')->names('post');
 
-Route::resource('admin/plan', PlanController::class)->names('plan');
+Route::post('admin/post/image', [PostController::class, 'updateImage'])->middleware('auth')->name('post.updateImage');
+
+Route::post('admin/plan/feature', [PlanController::class, 'addFeature'])->middleware('auth')->name('plan.addFeature');
+
+Route::delete('admin/plan/feature/{feature}', [PlanController::class, 'deleteFeature'])->middleware('auth')->name('plan.deleteFeature');
+
+Route::resource('admin/plan', PlanController::class)->middleware('auth')->names('plan');
 
 Auth::routes();
